huge file organization and consolidated server refactor
