+++ package.json
@@ -4,9 +4,7 @@
   "description": "service for users an valet information",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon account_server/index.js",
-    "login-server-dev": "nodemon login_server/index.js",
-    "server-dev": "nodemon index.js"
+    "start": "nodemon server/app.js"
   },
   "repository": {
     "type": "git",
@@ -24,8 +22,8 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "path": "^0.12.7",
-    "luxon": "^3.4.4"
+    "luxon": "^3.4.4",
+    "path": "^0.12.7"
   },
   "devDependencies": {
     "eslint": "^8.55.0",
