+++ server/db.js
@@ -1,0 +1,9 @@
+require('dotenv').config();
+const { createClient } = require('@supabase/supabase-js');
+
+const supabase = createClient(
+  process.env.Supabase_URL,
+  process.env.Supabase_API_Key
+);
+
+module.exports = supabase;
