+++ server/middleware/geocodeGrabber.js
@@ -1,4 +1,3 @@
-
 const { Client } = require('@googlemaps/google-maps-services-js');
 
 const google = new Client({});
@@ -10,34 +9,40 @@ const createGeocode = (address) => {
     const query = {
       params: {
         key: process.env.GOOGLE_KEY,
-        address
-      }
+        address,
+      },
     };
-    google.geocode(query)
-      .then(results => {
-        if (results.data.status !== 'OK') {
-          return reject('No results');
-        }
-        const coord = JSON.stringify(results.data.results[0].geometry.location);
-        return resolve(coord);
-      })
-  })
+    google.geocode(query).then((results) => {
+      if (results.data.status !== 'OK') {
+        return reject('No results');
+      }
+      const coord = JSON.stringify(results.data.results[0].geometry.location);
+      return resolve(coord);
+    });
+  });
 };
 
 const geocodeMiddleware = (req, res, next) => {
   //need to call the reservations server and add properties to the req.body object for the lat and lng based off of what is returned from the geocode api
   //build the whole address url
-  const address = req.body.address + ' ' + req.body.city + ' ' + req.body.state + ' ' + req.body.zip;
+  const address =
+    req.body.address +
+    ' ' +
+    req.body.city +
+    ' ' +
+    req.body.state +
+    ' ' +
+    req.body.zip;
   //call the google API with the address
   createGeocode(address)
-    .then(coord => {
+    .then((coord) => {
       //add the lat and lng properties to the req.body object
       const parsedCoord = JSON.parse(coord);
       req.body.lat = parsedCoord.lat;
       req.body.lng = parsedCoord.lng;
       next();
     })
-    .catch(err => {
+    .catch((err) => {
       console.error('Error creating geocode:', err);
       res.sendStatus(500);
     });
