+++ server/routes/loginRouter.js
@@ -1,10 +1,8 @@
 const express = require('express');
-const { createClient } = require('@supabase/supabase-js');
+
 const router = express.Router();
-
+const supabase = require('../db');
 const geocodeMiddleware = require('../middleware/geocodeGrabber');
-
-const supabase = createClient(process.env.Superbase_URL, process.env.Superbase_API_Key);
 
 //these are the routes that i will need to set up.
 //driver/valet account creation, get account info
@@ -35,65 +33,79 @@ router.get('/login/:id', async (req, res) => {
 //will probably refactor to be a supabase transaction for atomicity, so that if one fails, the other will fail as well. so we dont have to deal with a partially constructed record in the database.
 router.post('/login/:id/driver', async (req, res) => {
   const { id } = req.params;
-  const { google_accounts_id, first_name, last_name, email, phone_number, role, make, model, color, license_plate_number } = req.body;
+  const {
+    google_accounts_id,
+    first_name,
+    last_name,
+    email,
+    phone_number,
+    role,
+    make,
+    model,
+    color,
+    license_plate_number,
+  } = req.body;
   try {
-    const { error: accountsError } = await supabase
-      .from('accounts')
-      .insert(
-        {
-          id,
-          first_name,
-          last_name,
-          email,
-          phone_number,
-          role
-        }
-      );
+    const { error: accountsError } = await supabase.from('accounts').insert({
+      id,
+      first_name,
+      last_name,
+      email,
+      phone_number,
+      role,
+    });
     if (accountsError) {
       console.error('Error creating driver account:', accountsError);
       return res.sendStatus(500);
     }
-    const { error: carsError } = await supabase
-      .from('cars')
-      .insert(
-        {
-          make,
-          model,
-          color,
-          license_plate_number,
-          user_id: id
-        }
-      );
+    const { error: carsError } = await supabase.from('cars').insert({
+      make,
+      model,
+      color,
+      license_plate_number,
+      user_id: id,
+    });
     if (carsError) {
       console.error('Error creating driver account (cars):', carsError);
       return res.sendStatus(500);
     }
     res.sendStatus(201);
-  }
-  catch (error) {
+  } catch (error) {
     console.error('Error creating account:', error);
     res.status(500).json({ error: 'Internal server error' });
   }
-
 });
 //need to set up a post request to create a new valet account with the google id token and the user information that is input on the front end. this also assumes that a middleware function is used that generates the lat and lng from the passed in address. then creates the garage and the parking spots associated with that garage.
 router.post('/login/:id/valet', geocodeMiddleware, async (req, res) => {
   const { id } = req.params;
-  const { google_accounts_id, first_name, last_name, email, phone_number, role, address, city, state, zip, country, name, operation_hours, spots, lat, lng } = req.body;
+  const {
+    google_accounts_id,
+    first_name,
+    last_name,
+    email,
+    phone_number,
+    role,
+    address,
+    city,
+    state,
+    zip,
+    country,
+    name,
+    operation_hours,
+    spots,
+    lat,
+    lng,
+  } = req.body;
 
   try {
-    const { error: accountsError } = await supabase
-      .from('accounts')
-      .insert(
-        {
-          id,
-          first_name,
-          last_name,
-          email,
-          phone_number,
-          role
-        }
-      );
+    const { error: accountsError } = await supabase.from('accounts').insert({
+      id,
+      first_name,
+      last_name,
+      email,
+      phone_number,
+      role,
+    });
 
     if (accountsError) {
       console.error('Error creating valet account:', accountsError);
@@ -101,21 +113,19 @@ router.post('/login/:id/valet', geocodeMiddleware, async (req, res) => {
     }
     const { error: garagesError, data: garageData } = await supabase
       .from('garages')
-      .insert(
-        {
-          address,
-          city,
-          state,
-          zip,
-          country,
-          name,
-          operation_hours,
-          spots,
-          user_id: id,
-          lat,
-          lng
-        }
-      )
+      .insert({
+        address,
+        city,
+        state,
+        zip,
+        country,
+        name,
+        operation_hours,
+        spots,
+        user_id: id,
+        lat,
+        lng,
+      })
       .select();
 
     if (garagesError) {
@@ -123,10 +133,9 @@ router.post('/login/:id/valet', geocodeMiddleware, async (req, res) => {
       return res.sendStatus(500);
     }
 
-    const rows = Array.from({ length: spots }, () => (
-      {
-        garage_id: garageData[0].id
-      }));
+    const rows = Array.from({ length: spots }, () => ({
+      garage_id: garageData[0].id,
+    }));
 
     const { error: spotsError } = await supabase
       .from('parking_spots')
@@ -145,9 +154,7 @@ router.post('/login/:id/valet', geocodeMiddleware, async (req, res) => {
 
 router.get('/login/test/test', async (req, res) => {
   try {
-    const { data, error } = await supabase
-    .from('accounts')
-    .select('*');
+    const { data, error } = await supabase.from('accounts').select('*');
     //another valid way to query the data from this table would be to use the following:
     // const { data, error } = await supabase.from('accounts').select('*');
